+++ packages/web-runtime/src/defaults/languages.ts
@@ -1,0 +1,9 @@
+export const supportedLanguages = {
+  cs: 'Czech',
+  de: 'Deutsch',
+  en: 'English',
+  es: 'Español',
+  fr: 'Français',
+  gl: 'Galego',
+  it: 'Italiano'
+}
